+++ CHANGELOG.md
@@ -6,9 +6,11 @@
 - (feature) Add iterator, `entries`, `values`, and `keys`
 - (feature) Add `inspect`
 - (feature) Add `combine`
+
 * (feature) Static static constructors now use `Float64Array` when possible
 * (feature) `StridedView.random` now generates random values in the range `[0, 1)` passed to the `fill` function
 * (feature) Add `StridedView.cwise`
+* (feature) Add `neighbors` iterator
 
 ---